Move shared methods to base class
